requirements.txt
discord.py
pystyle